+++ packages/mcp/CHANGELOG.md
@@ -1,3 +1,24 @@
+# [@synergy-design-system/mcp-v1.17.2](https://github.com/synergy-design-system/synergy-design-system/compare/mcp/1.17.1...mcp/1.17.2) (2025-11-03)
+
+
+### Bug Fixes
+
+* 🐛 Async delimiter change for pre-selected value in syn-select not displayed ([#1061](https://github.com/synergy-design-system/synergy-design-system/issues/1061)) ([5f6d361](https://github.com/synergy-design-system/synergy-design-system/commit/5f6d36102b669cc34aa44cded47ef49559e74529))
+
+# [@synergy-design-system/mcp-v1.17.1](https://github.com/synergy-design-system/synergy-design-system/compare/mcp/1.17.0...mcp/1.17.1) (2025-11-03)
+
+
+### Bug Fixes
+
+* 🐛 Show scrollbars for closed sticky variant of side-nav ([#1062](https://github.com/synergy-design-system/synergy-design-system/issues/1062)) ([926905b](https://github.com/synergy-design-system/synergy-design-system/commit/926905bd15c87948700e5f32382547b2a1c43821))
+
+# [@synergy-design-system/mcp-v1.17.0](https://github.com/synergy-design-system/synergy-design-system/compare/mcp/1.16.0...mcp/1.17.0) (2025-11-03)
+
+
+### Features
+
+* ✨CD update for syn-alert ([#1064](https://github.com/synergy-design-system/synergy-design-system/issues/1064)) ([f9c0656](https://github.com/synergy-design-system/synergy-design-system/commit/f9c0656a268a71ea676ed0b937d8ca77cca2c924))
+
 # [@synergy-design-system/mcp-v1.16.0](https://github.com/synergy-design-system/synergy-design-system/compare/mcp/1.15.0...mcp/1.16.0) (2025-10-29)